Move test to RegistryManagerTests file
